fix copy to results dir

+++ protocols/CopyToResultsDir.sh
@@ -162,14 +162,16 @@ do
 		printf "."
 	fi
 	echo "copying Convading data"
-	cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.totallist.txt ${projectResultsDir}/variants/cnv/
-	cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.shortlist.txt ${projectResultsDir}/variants/cnv/
-	cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.longlist.txt ${projectResultsDir}/variants/cnv/
-	cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.log ${projectResultsDir}/variants/cnv/
-	cp ${intermediateDir}/Convading//CreateFinalList/${sa}/*.shortlist.finallist.txt ${projectResultsDir}/variants/cnv/
-	echo "copying XHMM results"
-	cp ${intermediateDir}/${sa}_step10.xcnv ${projectResultsDir}/variants/cnv/
-	
+	if [ -f ${intermediateDir}/${sa}_step10.xcnv ]
+	then
+		cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.totallist.txt ${projectResultsDir}/variants/cnv/
+		cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.shortlist.txt ${projectResultsDir}/variants/cnv/
+		cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.longlist.txt ${projectResultsDir}/variants/cnv/
+		cp ${intermediateDir}/Convading//StartWithBestScore/${sa}/*.only.best.score.log ${projectResultsDir}/variants/cnv/
+		cp ${intermediateDir}/Convading//CreateFinalList/${sa}/*.shortlist.finallist.txt ${projectResultsDir}/variants/cnv/
+		echo "copying XHMM results"
+		cp ${intermediateDir}/${sa}_step10.xcnv ${projectResultsDir}/variants/cnv/
+	fi
 done
 printf " finished (7/11)\n"
